Added support for subscribing to stream for ESDB Consumer
